[ALPS03975965] WLAN: fix coverity Printf arg type mismatch
[Detail]
fix coverity Printf arg type mismatch
cast unsigned long to uint64_t and change %u to %lu

MTK-Commit-Id: 52cd174b3a3c427268292f093a7189d83d6aec33

CR-Id: ALPS03975965
Feature: WIFI
(cherry picked from commit d525362c7d2df024504e55c188523416d9ea2433)

+++ os/linux/hif/common/hal_pdma.c
@@ -1138,8 +1138,8 @@ bool halWpdmaAllocTxRing(struct GLUE_INFO *prGlueInfo, uint32_t u4Num,
 		return false;
 	}
 
-	DBGLOG(HAL, TRACE, "TxDescRing[%p]: total %u bytes allocated\n",
-	       prTxDesc->AllocVa, prTxDesc->AllocSize);
+	DBGLOG(HAL, TRACE, "TxDescRing[%p]: total %lu bytes allocated\n",
+	       prTxDesc->AllocVa, (uint64_t)prTxDesc->AllocSize);
 
 	/* Save PA & VA for further operation */
 	RingBasePa = prTxDesc->AllocPa;
@@ -1203,8 +1203,8 @@ bool halWpdmaAllocRxRing(struct GLUE_INFO *prGlueInfo, uint32_t u4Num,
 		return false;
 	}
 
-	DBGLOG(HAL, TRACE, "RxDescRing[%p]: total %u bytes allocated\n",
-		prRxDesc->AllocVa, prRxDesc->AllocSize);
+	DBGLOG(HAL, TRACE, "RxDescRing[%p]: total %lu bytes allocated\n",
+		prRxDesc->AllocVa, (uint64_t)prRxDesc->AllocSize);
 
 	/* Initialize Rx Ring and associated buffer memory */
 	RingBasePa = prRxDesc->AllocPa;
